+++ scripts/provision/debian/custom/saltstack_formulas/A_base.sh
@@ -6,8 +6,6 @@ echo '##############################################'
 echo "Starting ${0}.."
 set -x
 
-minion_config_path=/vagrant/share/salt-config/${HOSTNAME}/config/minion
-master_config_path=/vagrant/share/salt-config/${HOSTNAME}/config/master
 states_top_path=/vagrant/share/salt-config/${HOSTNAME}/file_roots/states/top.sls
 pillar_root=/vagrant/share/salt-config/${HOSTNAME}/file_roots/pillar/
 
@@ -15,8 +13,6 @@ mkdir -p /srv/salt/{_grains,_modules/formulas,_states,contrib/states,pillar/exam
 
 [[ -f $states_top_path ]] && ln -sf $states_top_path /srv/salt/states/top.sls
 [[ -d $pillar_root && ! -e /srv/salt/pillar/share ]] && ln -sf $pillar_root /srv/salt/pillar/share
-[[ -f $minion_config_path ]] && cp $minion_config_path /etc/salt/
-[[ -f $master_config_path ]] && cp $master_config_path /etc/salt/
 
 if [[ -d /vagrant/salt/formulas/ ]]; then
   for d in /vagrant/salt/formulas/*; do