+++ CHANGELOG.rst
@@ -1,3 +1,4 @@
 [Unreleased] ([unreleased])
 ---------------------------
+* improve provision scripts
 * initial version